fix(ci): file-size guard — extract workspace overview cursor pagination
