+++ src/awf/service/workspace_observability.py
@@ -3,9 +3,6 @@
 from __future__ import annotations
 
 import asyncio
-import base64
-import binascii
-import json
 import re
 from collections.abc import Iterator, Mapping, Sequence
 from contextlib import contextmanager
@@ -51,6 +48,18 @@ from awf.service.usage_store import (
     UsageSnapshot,
     read_latest_usage_snapshot,
     read_latest_usage_snapshots,
+)
+from awf.service.workspace_overview_pagination import (
+    InvalidWorkspaceOverviewCursorError as InvalidWorkspaceOverviewCursorError,
+)
+from awf.service.workspace_overview_pagination import (
+    _decode_overview_cursor as _decode_overview_cursor,
+)
+from awf.service.workspace_overview_pagination import (
+    _encode_overview_cursor as _encode_overview_cursor,
+)
+from awf.service.workspace_overview_pagination import (
+    _WorkspaceOverviewCursor as _WorkspaceOverviewCursor,
 )
 
 AgentIdentitySource = Literal["task_policy", "default", "unavailable"]
@@ -128,16 +137,6 @@ _GENERIC_RECOVERY_REASON_CODES = frozenset(
 _MAX_RECOVERY_PAYLOAD_KEYS = 32
 _MAX_RECOVERY_PAYLOAD_DEPTH = 4
 _MAX_RECOVERY_PAYLOAD_SEQUENCE_ITEMS = 20
-
-
-@dataclass(frozen=True)
-class _WorkspaceOverviewCursor:
-    created_at: datetime
-    workspace_id: str
-
-
-class InvalidWorkspaceOverviewCursorError(ValueError):
-    """Raised when a workspace overview pagination cursor cannot be decoded."""
 
 
 @dataclass(frozen=True)
@@ -513,40 +512,6 @@ def _provider_readiness_preflight_from_task_policy(
         return None
     value = task_policy.get("provider_readiness_preflight")
     return dict(value) if isinstance(value, Mapping) else None
-
-
-def _encode_overview_cursor(workspace: Workspace) -> str:
-    payload = {
-        "t": workspace.created_at.isoformat(),
-        "id": workspace.id,
-    }
-    encoded = base64.urlsafe_b64encode(json.dumps(payload, separators=(",", ":")).encode("utf-8"))
-    return encoded.decode("ascii").rstrip("=")
-
-
-def _decode_overview_cursor(cursor: str | None) -> _WorkspaceOverviewCursor | None:
-    if cursor is None:
-        return None
-    try:
-        padded_cursor = cursor + ("=" * (-len(cursor) % 4))
-        decoded = base64.urlsafe_b64decode(padded_cursor.encode("ascii"))
-        payload = json.loads(decoded.decode("utf-8"))
-        created_at = datetime.fromisoformat(
-            payload["t"] if "t" in payload else payload["created_at"]
-        )
-        workspace_id = payload["id"] if "id" in payload else payload["workspace_id"]
-    except (
-        binascii.Error,
-        KeyError,
-        TypeError,
-        UnicodeDecodeError,
-        ValueError,
-        json.JSONDecodeError,
-    ) as exc:
-        raise InvalidWorkspaceOverviewCursorError("Invalid workspace overview cursor") from exc
-    if not isinstance(workspace_id, str) or workspace_id == "":
-        raise InvalidWorkspaceOverviewCursorError("Invalid workspace overview cursor")
-    return _WorkspaceOverviewCursor(created_at=created_at, workspace_id=workspace_id)
 
 
 def effective_agent_identity(
